# Regex set optimisation when looking series in ingester