--- a/pkg/ingester/index/index_test.go
+++ b/pkg/ingester/index/index_test.go
@@ -1,6 +1,9 @@
 package index
 
 import (
+	"fmt"
+	"strconv"
+	"strings"
 	"testing"
 
 	"github.com/prometheus/common/model"
@@ -61,6 +64,67 @@ func TestIndex(t *testing.T) {
 	assert.Equal(t, []string{"flip", "foo"}, index.LabelNames())
 	assert.Equal(t, []string{"bar", "baz"}, index.LabelValues("foo"))
 	assert.Equal(t, []string{"flap", "flop"}, index.LabelValues("flip"))
+}
+
+func BenchmarkSetRegexLookup(b *testing.B) {
+	// Prepare the benchmark.
+	seriesLabels := []string{"a", "b", "c", "d", "e", "f", "g", "h", "i", "j"}
+	seriesPerLabel := 100000
+
+	idx := New()
+	for _, l := range seriesLabels {
+		for i := 0; i < seriesPerLabel; i++ {
+			lbls := labels.FromStrings("foo", l, "bar", strconv.Itoa(i))
+			idx.Add(client.FromLabelsToLabelAdapters(lbls), model.Fingerprint(lbls.Hash()))
+		}
+	}
+
+	selectionLabels := []string{}
+	for i := 0; i < 100; i++ {
+		selectionLabels = append(selectionLabels, strconv.Itoa(i))
+	}
+
+	tests := []struct {
+		name    string
+		matcher string
+	}{
+		{
+			name:    "select all",
+			matcher: fmt.Sprintf(`{bar=~"%s"}`, strings.Join(selectionLabels, "|")),
+		},
+		{
+			name:    "select two",
+			matcher: fmt.Sprintf(`{bar=~"%s"}`, strings.Join(selectionLabels[:2], "|")),
+		},
+		{
+			name:    "select half",
+			matcher: fmt.Sprintf(`{bar=~"%s"}`, strings.Join(selectionLabels[:len(selectionLabels)/2], "|")),
+		},
+		{
+			name:    "select none",
+			matcher: `{bar=~"bleep|bloop"}`,
+		},
+		{
+			name:    "equality matcher",
+			matcher: `{bar="1"}`,
+		},
+		{
+			name:    "regex (non-set) matcher",
+			matcher: `{bar=~"1.*"}`,
+		},
+	}
+
+	b.ResetTimer()
+
+	for _, tc := range tests {
+		b.Run(fmt.Sprintf("%s:%s", tc.name, tc.matcher), func(b *testing.B) {
+			matcher := mustParseMatcher(tc.matcher)
+			for n := 0; n < b.N; n++ {
+				idx.Lookup(matcher)
+			}
+		})
+	}
+
 }
 
 func mustParseMatcher(s string) []*labels.Matcher {