@cyriltovena 2020-04-16
Also closes #1993