@pstibrany 2020-04-16
> Changelog entry is missing

Perhaps can just extend entry for  #2446, up to you.